Allow enabling "Bundler 3 mode" more easily
Currently to test Bundler 3 mode we have to actually edit the version
file to simulate we're running a future version. This is inconvenient.

Instead, allow passing an environment variable,
`BUNDLER_EFFECTIVE_VERSION`, to set the "working mode" Bundler should
use.

This can now be set easily by end users to enable them to try out the
changes in the future version and give us feedback.

There's is one small change with respect to the previous way. When running
something like `bundle update --bundler 2.3.0`, what Bundler does is
install Bundler 2.3.0 and then re'exec using that. Previously, the
version file would of course not be changed in the 2.3.0 version just
installed, so "the simulated future Bundler" would start behaving as the
standard Bundler after re'exec. Now, since the behavior is controlled by
an ENV variable, you also get "Bundler 3 mode" after re'excing. I think
this is fine and I only mention it because I needed to change a couple
of specs due to this change (I added the `--verbose` flag to them so
that the output they check is not dependant on the version of Bundler,
since Bundler 3 is less verbose that Bundler 2 by default).